add common picker module for logic deduplication
