+++ lua/laravel/pickers/telescope/preview.lua
