+++ lua/laravel/pickers/ui_select/actions.lua
@@ -1,0 +1,10 @@
+local common_actions = require("laravel.pickers.common.actions")
+local ui_run = require("laravel.pickers.ui_select.ui_run")
+
+local M = {}
+
+function M.run(command)
+  common_actions.run(command, ui_run)
+end
+
+return M
